+++ src/cli/migrations/index.js
@@ -13,6 +13,7 @@ import * as m0021 from './0.0.21.js'
 import * as m0022 from './0.0.22.js'
 import * as m0023 from './0.0.23.js'
 import * as m0024 from './0.0.24.js'
+import * as m0025 from './0.0.25.js'
 
 /** Migrations sorted by version (ascending). */
-export const migrations = [m0010, m0016, m0017, m0018, m0019, m0020, m0021, m0022, m0023, m0024]
+export const migrations = [m0010, m0016, m0017, m0018, m0019, m0020, m0021, m0022, m0023, m0024, m0025]